Working my remote ctags.nvim plugin

+++ vim-utils/rplugin/python3/ctags.py
@@ -2,7 +2,7 @@
 # Description: Initializer for all my remote plugins
 # Author:Reinaldo Molina <rmolin88@gmail.com>
 # Version:1.0.0
-# Last Modified: Mar 15 2017 10:13
+# Last Modified: Thu Mar 16 2017 16:35
 # Created: Mar 15 2017 10:13
 
 import os
@@ -18,21 +18,27 @@ class RemovePlugin(object):
     @neovim.function('UpdateTagsRemote')
     def update_tags_remote(self, args):
         if self.busy != 0:
-            #  print("UpdateTagsRemote() is busy")
-            self.nvim.command(':echomsg "UpdateTagsRemote() is busy"')
             return
-
         self.busy = 1
 
-        delete_files = [ 'cscope.files', 'cscope.out', 'cscope.po.out', 'cscope.in.out', '.tags' ]
-        file_ext_tuple = ('.c', '.cpp', '.java', '.cc', '.h', '.hpp')
-        ctags_cmd = 'ctags -R -L cscope.files -f .tags --sort=yes --c++-kinds=+pl --fields=+iaS --extra=+q --language-force=C++'
+        DELETED_FILES = ( 'cscope.files', 'cscope.out', 'cscope.po.out', 'cscope.in.out', '.tags' )
+        #  EXT = ('.c', '.cpp', '.java', '.cc', '.h', '.hpp')
+
+        # TODO.RM-Thu Mar 16 2017 17:17: Put the tag in the cache folder along with the cscope
+        lang = self.nvim_ft_to_ctags()
+        if lang is None:
+            self.nvim.command(':echomsg "Failed to retrieve &filetype"')
+            self.busy = 0
+            return
+
+        CMD = 'ctags -R -L cscope.files -f .tags --sort=yes --c++-kinds=+pl --fields=+iaS \
+                --extra=+q --language-force=%s' % lang
 
         #  Delete current connection if there is one
         self.nvim.command('silent! cs kill -1')
 
         # Silently delete files
-        for files in delete_files:
+        for files in DELETED_FILES:
             try:
                 os.remove(files)
             except:
@@ -41,10 +47,10 @@ class RemovePlugin(object):
         # Populate list of source files
         #  list_files = open('cscope.files', 'w+')
         os.system('rg --files -t cpp . > cscope.files')
-        #  for root, dirs, files in os.walk(os.getcwd()):
-            #  for file in files:
-                #  if file.endswith(file_ext_tuple):
-                    #  list_files.write(os.path.join(root, file) + "\n")
+        if os.path.isfile('cscope.files') == False:
+            self.nvim.command(':echomsg "Failed to create cscope.files"')
+            self.busy = 0
+            return
 
         # Silently try to create cscope files
         try:
@@ -55,31 +61,35 @@ class RemovePlugin(object):
         # Check that cscope was created if so. Silently create tags
         if os.path.isfile('cscope.out'):
             try:
-                os.system(ctags_cmd)
+                os.system(CMD)
             except:
                 pass
+        else:
+            self.nvim.command(':echomsg "Failed to create cscope.out"')
+            self.busy = 0
+            return
 
         # Add new database
         self.nvim.command('cs add cscope.out')
         self.busy = 0
 
-#  def _increment_calls(self):
-    #  if self.calls == 5:
-        #  raise Exception('Too many calls!')
-            #  self.calls += 1
+    def nvim_ft_to_ctags(self):
+        #  Determine for what language is the tags going to be
+        lang = self.nvim.eval('&ft').lower().split('.')
+        if lang is None:
+            return []
 
-#  @neovim.command('Cmd', range='', nargs='*', sync=True)
-#  def command_handler(self, args, range):
-    #  self._increment_calls()
-    #  self.nvim.current.line = (
-        #  'Command: Called %d times, args: %s, range: %s' % (self.calls,
-                                                           #  args,
-                                                           #  range))
+        if lang[0] == 'cpp':
+            lang[0] = 'C++'
+        elif lang[0] == 'vim':
+            lang[0] = 'Vim'
+        elif lang[0] == 'python':
+            lang[0] = 'Python'
+        elif lang[0] == 'java':
+            lang[0] = 'Java'
+        else:
+            return []
 
-#  @neovim.autocmd('BufEnter', pattern='*.py', eval='expand("<afile>")',
-                #  sync=True)
-#  def autocmd_handler(self, filename):
-    #  self._increment_calls()
-    #  self.nvim.current.line = (
-        #  'Autocmd: Called %s times, file: %s' % (self.calls, filename))
-
+        #  Use for debugging        
+        #  self.nvim.command(':echomsg "lang = %s"' %lang[0])
+        return lang[0]
